requirements.txt
paddlepaddle==2.6.1
opencv-python==4.6.0.66
pdf2image==1.17.0
numpy==1.26.4
Pillow==10.4.0
tqdm==4.66.4

# Dépendances runtime de paddleocr (sans PyMuPDF)
opencv-contrib-python==4.6.0.66
shapely==2.1.1
scikit-image==0.24.0
imgaug==0.4.0
pyclipper==1.3.0.post6
lmdb==1.7.3
visualdl==2.5.3
rapidfuzz==3.14.1
lxml==6.0.1
premailer==3.10.0
openpyxl==3.1.5
attrdict==2.0.1
pyyaml==6.0.2
python-docx==1.2.0
beautifulsoup4==4.13.5
fonttools==4.59.2
fire==0.7.1
pdf2docx==0.5.8
